+++ src/Application/Interface.ts
@@ -6,7 +6,6 @@
  * @includes ICookie Application/_Interface/ICookie
  * @includes IConfig Application/_Interface/IConfig
  * @includes IEnv Application/_Interface/IEnv
- * @includes ILocation Application/_Interface/ILocation
  * @includes ISerializableState Application/_Interface/ISerializableState
  * @includes IStore Application/_Interface/IStore
  * @includes IRequest Application/_Interface/IRequest
@@ -21,7 +20,7 @@
 export * from 'Application/_Interface/ICookie';
 export * from 'Application/_Interface/IConfig';
 export * from 'Application/_Interface/IEnv';
-export * from 'Application/_Interface/ILocation';
+// export * from 'Application/_Interface/ILocation';
 export * from 'Application/_Interface/ISerializableState';
 export * from 'Application/_Interface/IStore';
 // export * from 'Application/_Interface/IStateReceiver';
