+++ src/Application/_Env/NodeJS/Location.ts
@@ -1,5 +1,6 @@
 /// <amd-module name="Application/_Env/NodeJS/Location" />
-import { ILocation, IHttpRequest } from 'Application/Interface';
+import { IHttpRequest } from 'Application/Interface';
+import type { ILocation } from 'Application/_Env/Interfaces;
 
 export default class Location implements ILocation {
     private hostMask: RegExp = /([^:]+)(:\d+)?/;
